add Root component for hot loading
